+++ tests/ui/unpretty/interpolation-expanded.rs
