+++ tests/ui/unpretty/interpolation-expanded.stdout
